+++ common/src/main/java/io/github/xiewuzhiying/vs_addition/mixinducks/valkyrienskies/EntityDraggingInformationMixinDuck.java
@@ -1,0 +1,7 @@
+package io.github.xiewuzhiying.vs_addition.mixinducks.valkyrienskies;
+
+public interface EntityDraggingInformationMixinDuck {
+    double addedPitchRotLastTick = 0.0;
+    double getAddedPitchRotLastTick();
+    void setAddedPitchRotLastTick(double d);
+}