+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Presence footsteps compat (I don't understand why only carpets and pressure plates etc. work🤔)
 - Basically move to kotlin
 - Fix confusing PSI mode
+- Extended entity dragger
 ## Fabric
 - Fix the issue of crashing caused by not including mixin squared
 ## Forge